[all] MessageReceiver: Implement extended messages, move to common/

+++ common/include/MessageReceiver.hpp
@@ -34,8 +34,8 @@ protected:
 	void handleIncomingData(const QByteArray &readData);
 
 	RxStatus m_rxStatus = MSG_RX_MAGIC;
-	quint16 m_rxByteCount = 0;
-	quint16 m_rxMsgSize = 0;
+	quint32 m_rxByteCount = 0;
+	quint32 m_rxMsgSize = 0;
 	quint16 m_rxMsgID = 0;
 	QByteArray m_rxBuffer;
 };
